requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
DateTime==4.3
Flask==2.0.1
Flask-Assets==2.0
Flask-JWT-Extended==4.2.1
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
webassets==2.0
Werkzeug==2.0.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
